RIGID_BODY!
Ok, rearanged things (as always) and right now I have a standard rigid body scene. Mesh incase I need to see it, and just a square collision.

The parent (root of the scene) has a rigid body script that will look at all of its children and move them to the actual rigid bodies position.

I did this because putting the actual elements I want following as children of the rigidBody3D, then it makes it difficult to move it in other ways. Plus it just gets cluttered and they dont REALLY belong to the rigidBody3d, so next best thing felt like putting it as a child of the scene. I won't even have to edit as children.

Also figured out generally how I want to handle interactions via collisions. Created a collision_body scene which gives a basic structure with an Area3D